# openapi.yaml
openapi: 3.1.0
info:
  title: DOGFOOD Portal API
  version: 1.0.0
  description: >-
    Local-first hackathon submission, judging, community, and organizer API.
    Session identity is carried in an HttpOnly session cookie. Runtime data is in memory.
servers:
  - url: http://localhost:8080
security: []
paths:
...
  /api/v1/projects:
    get:
      summary: Search public projects
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: track, in: query, schema: { type: string } }
      responses:
        "200": { description: Filtered project list }
    post:
      summary: Submit a project for the participant's team
      security: [{ SessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201": { description: Project created }
        "403": { description: Participant role or open deadline required }
  /api/v1/projects/{id}:
    get:
      summary: Read a public project and community data
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200": { description: Project detail; judge results remain private before publication }
        "404": { description: Project not found }
    put:
      summary: Edit the participant's project before the deadline
      security: [{ SessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "200": { description: Project updated }
        "403": { description: Team ownership or deadline check failed }
  /api/v1/projects/{id}/votes:
    post:
      summary: Cast one rate-limited community vote
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email, description: Required when event voting access is email-gated }
      responses:
        "201": { description: Vote recorded }
        "409": { description: Duplicate vote }
        "429": { description: Rate limit exceeded }
  /api/v1/projects/{id}/comments:
    post:
      summary: Add a rate-limited public comment
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                author: { type: string, maxLength: 80 }
                text: { type: string, maxLength: 2000 }
      responses:
        "201": { description: Comment recorded }
        "429": { description: Rate limit exceeded }
  /api/v1/judge/scores:
    post:
      summary: Submit a private ballot for an assigned project
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project, criteria]
              properties:
                project: { type: string }
                criteria: { type: object, additionalProperties: { type: integer, minimum: 1, maximum: 5 } }
                comment: { type: string, maxLength: 4000 }
      responses:
        "201": { description: Ballot recorded }
        "403": { description: Judge role or assignment required }
...
  /api/v1/export/projects.csv:
    get:
      summary: Export project submissions as CSV
      security: [{ SessionCookie: [] }]
      responses:
        "200": { description: CSV project export, content: { text/csv: {} } }
  /api/v1/import:
    post:
      summary: Bulk import up to 500 projects
      security: [{ SessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [projects]
              properties:
                projects: { type: array, maxItems: 500, items: { $ref: "#/components/schemas/ProjectInput" } }
      responses:
        "201": { description: Imported projects }
        "400": { description: Invalid team, track, or project data }
  /api/v1/webhooks:
    post:
      summary: Register an event webhook endpoint
      security: [{ SessionCookie: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [url, event]
              properties:
                url: { type: string, format: uri }
                event: { type: string, enum: ["*", project.submitted, project.updated, project.imported, score.recorded, vote.cast, comment.created, team.created, team.invited, team.joined, event.updated, judge.invited, judge.joined, judging.assignments.created, judging.rubric.updated, certificate.generated] }
      responses:
        "201": { description: Webhook registration created; response includes the HMAC secret }
...
  /api/v1/judges/{judgeId}/record:
    get:
      summary: Generate a signed judge participation record
      parameters:
        - { name: judgeId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Ed25519-signed record and public key }
  /api/v1/judges/records/verify:
    post:
      summary: Verify a signed participation record's integrity
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: Signature integrity result }
  /api/certificates/{teamId}:
    get:
      summary: Generate a plain-text team certificate
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Certificate text, content: { text/plain: {} } }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    ProjectInput:
      type: object
      required: [title]
      properties:
        title: { type: string, maxLength: 120 }
        track: { type: string }
        summary: { type: string, maxLength: 2000 }
        repo_url: { type: string, format: uri }